+++ src/pnpjs/Lists/getList/getSourceList.ts
@@ -5,7 +5,7 @@ import { IFpsErrorObject } from '../../Common/IFpsErrorObject';
 import { checkAnyResults } from '../../Common/CheckAnyResults';
 
 /**
- * getSourceItems calls the Pnp function to get the results which returns the raw error.
+ * getSourceList calls the Pnp function to get the results which returns the raw error.
  * This function then will convert the error into the helpful error and return the standard IItemsError object.
  * 
  * @param sourceProps