+++ CHANGELOG.md
@@ -14,6 +14,7 @@ NOTE:  TURN BACK ON STRICT NULL CHECKS - turned off for PropPaneCols
 ## 1.0.80 - 2022-Dec-23: clean up console.logs
 - remove unneccessary console.logs
 - fix a couple check4Gulp imports that pull in pnp2
+- update PinMeGroup to only auto-expand on PageInfo for now
 
 ## 1.0.79 - 2022-Dec-23: fetch attacments pnpv2 update
 update fetch attachments from issue:  https://github.com/mikezimm/drilldown7/issues/302